refactor(discounts): make PromotionSystem class file
