+++ __tests__/PromotionSystemTest.js
@@ -1,22 +1,7 @@
 import DateUtils from "../src/utils/DateUtils";
 import StockSystem from "../src/inventory/StockSystem.js";
+import PromotionSystem from "../src/discounts/PromotionSystem";
 import path from "path";
-
-class PromotionSystem {
-  activePromotions() {
-    const today = DateUtils.getKoreaNow();
-    const currentKoreaFormattedDate = DateUtils.getKoreaNowFormatted(today);
-    const promtionInfoList = StockSystem.parseFile("promotions.md");
-    const activePromotions = [];
-    promtionInfoList.forEach((promotion) => {
-      if (promotion.start_date <= currentKoreaFormattedDate && promotion.end_date >= currentKoreaFormattedDate) {
-        activePromotions.push(promotion.name);
-      }
-    });
-
-    return activePromotions;
-  }
-}
 
 describe("PromotionSystem 클래스", () => {
   test("입력받은 오늘이 반짝할인, MD추천상품, 탄산2+1 프로모션이 가능한날짜인지 확인", () => {
